handle NULL CommandSource & other fixes
