+++ patches/api/0480-Brigadier-based-command-API.patch
@@ -80,10 +80,10 @@ index 0000000000000000000000000000000000000000..7e24babf746de474c8deec4b147e2203
 +}
 diff --git a/src/main/java/io/papermc/paper/command/brigadier/CommandSourceStack.java b/src/main/java/io/papermc/paper/command/brigadier/CommandSourceStack.java
 new file mode 100644
-index 0000000000000000000000000000000000000000..f7685aeef9d7306d8b2f7e62315b99acc87d9202
+index 0000000000000000000000000000000000000000..54288dbe7185b875a74184f002ee4de4405e91b1
 --- /dev/null
 +++ b/src/main/java/io/papermc/paper/command/brigadier/CommandSourceStack.java
-@@ -0,0 +1,46 @@
+@@ -0,0 +1,50 @@
 +package io.papermc.paper.command.brigadier;
 +
 +import org.bukkit.Location;
@@ -100,6 +100,10 @@ index 0000000000000000000000000000000000000000..f7685aeef9d7306d8b2f7e62315b99ac
 + * important additional context for the command execution.
 + * Specifically, commands such as {@literal /execute} may alter the location or executor of the source stack before
 + * passing it to another command.
++ * <p>The {@link CommandSender} returned by {@link #getSender()} may be a "no-op"
++ * instance of {@link CommandSender} in cases where the server either doesn't
++ * exist yet, or no specific sender is available. Methods on such a {@link CommandSender}
++ * will either have no effect or throw an {@link UnsupportedOperationException}.</p>
 + */
 +@ApiStatus.NonExtendable
 +@ApiStatus.Experimental
