+++ test-plugin/src/main/java/io/papermc/testplugin/brigtests/example/IceCreamType.java
@@ -1,4 +1,4 @@
-package io.papermc.testplugin.example;
+package io.papermc.testplugin.brigtests.example;
 
 public enum IceCreamType {
     VANILLA,
